type and printf specifier fixes for i386
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@10028 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ share/cfg_common.c
@@ -296,7 +296,7 @@ slcfg_getif(struct ifconf *ifc, struct addrinfo *ai, char ifn[IFNAMSIZ])
 
 	errno = 0;
 	rc = write(s, &rq, rq.nmh.nlmsg_len);
-	if (rc != rq.nmh.nlmsg_len)
+	if (rc != (ssize_t)rq.nmh.nlmsg_len)
 		psc_fatal("routing socket length mismatch");
 
 	rc = read(s, &rq, sizeof(rq));
